requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.5
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.1
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
whitenoise==5.2.0
